+++ RELEASE_NOTES.rst
@@ -1,6 +1,12 @@
 Sphinx Repo Manager Release Notes
 =================================
 
+v1.0.19
+-------
+
+* Injected auth via headers instead of URL params for git operations for a more-sanitized approach
+* Added ``tools/admin-enable-long-file-paths.ps1``
+
 v1.0.18
 -------